(chore) move PostReview from subfolder to components/Review
