+++ frontend/src/components/review/ReviewPage.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useMemo } from "react";
 import { useRecoilValue, useResetRecoilState } from "recoil";
 import { reviewSettingsState, reviewStageState } from "state/atoms/reviewAtoms";
-import PostReview from "./PostReview/PostReview";
+import PostReview from "./PostReview";
 import PreReview from "./PreReview/PreReview";
 import Review from "./Review";
